Refactor VideoMetadata
